Added function stubs
