+++ TrainOfThought/TrainOfThought/urls.py
@@ -16,10 +16,11 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
-from . import views
+from backend import views
 
 urlpatterns = [
     path('', views.index, name='index'),
     path('admin/', admin.site.urls),
     path('gpt-post', views.gpt_post, name='gpt-post'),
+    path('tick_posts/', views.tick_posts, name='tick_data'),
 ]
